+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [1.0.1] - 2024-03-01
+### Updated
+- Fixed undefined variable $meta_value.
+- Tested compatible with PHP 7.4 to 8.2.
+- Tested compatible with WordPress up to version 6.4.3.
+- Updated copyright notice for 2024.
+
 ## [1.0.0] - 2022-01-06
 ### Updated
 - Release 1.0